fix issue described
ISSUE: typing in an input created by Form caused a loss of focus
so that only 1 character could be typed

CAUSE: the key of the Input created by Form was crypto.randomUUID().
This meant that when react re-rendered the App, which happened when
a character was typed into an input, each Input would have a
different, random key generated. This means that react would not
recognise each Input as the same as the previous version of itself,
and so would cause a loss of focus.
The Section component also gave each Form component a random key,
leading to the same problem.

FIX: Change Form and Section so that they pass an unchanging key
when creating components